Added code for the receiving end
Here, I implemented the other side of the websocket. The monitor is
registered as a client as well on the websocket server. But on
connection it receives a list of available cameras to display streams
from. The list is communicated as individual messages as this works for
the initial connection as well as cameras joining later on.