Fixed the page being scrollable
Although my content is positioned absolutely and is not bigger than the
viewport, I still had a scrollbar. I dislike scrollbars that are there
for no reason and wanted to remove it. The issue was some left over
margin on the body from the user agent styles in chrome.

With this simple CSS rule, the scrollbar is gone.