Full multi-cam support
Here I improved what I started building in the last commit. With the
websocket communication working, I could upgrade the frontend to allow
for multiple cameras to be monitored at once.

I extracted the actual display of the video feed and the controls into
a separate Screen component. This still needs a flag for wether it's
receiving or sending.

On the monitor side cameras are now connected and disconnected on the
fly which makes for a great user experience in my opinion.