Separate UIs for sending and receiving end
In this commit I implemented a separate ui for the receiving end as
there, I think we don't need controls to turn the camera off. But I
think turning off the audio is something that I want to be able to when
monitoring my daughter.

This also serves as a solution for autoplay. I can only autoplay videos
muted, so when loading the monitor page, all videos start muted now.