Implemented meta-data communication
In this commit I implemented communication of meta-data from the camera
to all connected monitors.

This allows me to send the battery and charging status to the monitor. I
wanted to add this feature, so that I can see on the monitor whether the
camera is running out of battery.