Implemented docker setup (for development)
Originally I planned on integrating the websocket server into the
sveltekit app. But using bun this is not straight forward and would
require some workarounds in order to be able to use the bun server in
development.

I realized, I don't need to force those two systems together, when I
can host them individually with an nginx as a reverse proxy in front of
them.

This is what I did here. I'll see how well this works. Eventually, I'll
create a setting for defining the hostname of the websocket server in
the sveltekit app. This creates maximum flexibility.

I could rewrite the whole thing and ditch sveltekit completely, serving
svelte pages via bun. But that's more work than I currently want to put
in. I'd rather spent my time getting this app to a usable state.