Fixed error setting meta-data when camera connects after monitor
Each monitor keeps track of it's cameras in a map. I made the mistake of
putting a newly connected camera into this map only after everything
else was setup. Including the async offer creation.
So it could happen that a meta-data message would arrive, when the
camera was in the process of being setup and not yet stored in the map.

An error would be thrown as I was trying to set meta-data to an
undefined value.

Here, I fixed this by stoding the new camera in the map right after
creation.