[WIP] Started rewrite
With the proof of concept working an countless hours of trying to
refactor and abstract it unsuccessful, I decided to start fresh.

Here, I implemented a simple symmetric communication procedure to
connect clients with each other.

This allows me to remove state from the
server - the server now only tracks open connections from clients and
closes duplicate ones - a necessity due to HMR creating new connections
with each code update.

I also implemented a simple scheme to reestablish lost connections - a
feature that the proof of concept lacked.

I used no abstractions at all, as I first want to abstract only when it
becomes too painful not to do so - or when I see the bigger picture and
a good path to abstraction.