Rewrite completed
After a lot of fiddling, I finally got the next, more robust version of
Yuzukam to work.

We now have the same code for each peer with an introduction system, so
the server does no longer need to keep track of the state.

Also, connections are more stable, server restarts, page reloads of one
peer, do not break the connection. It is just reestablished.

This is really jusy a proof of concept. In the next commits, I'll make
this app more usable.