Basic functionality and typing
Here I added the base functionality that I envisioned:
- Users can name their client. This helps identify clients when
  multiple are connected.
- Users can switch between breadcasating their camera stream and just
  receiving other client streams. I've managed to do it in a way that
  the MediaStream is completely terminated and the browser also shows
  you, that no camera or mic is being used. This makes for better UX and
  saves some ressources on the clients device.
- The clients metadata (for now only the name and type) are send to
  other peers. Battery is not yet included in this. I'll add this back
  later. I decided agaist sending the metadata over the
  RTCPeerConnection as this would require a data channel and more code
  to make it work.

Then, I also tried to type everything correctly. And I mostly made it!
There are a few places, where it was too complicated to find a
specifically typed solution, and I fell back to a generically typed
apporach. But all APIs are typed in a way, that you get correct
suggestions and error messages! I am happy about this, although it was
quite some work. Makes using the code more convenient.