Simplified and more robust hosting with custom Docker image
Here, I adapted the code so that this app can be hosted with a single
docker image, as opposed to using three like how I do it when developing.

I also made sure to build all code so that this image is more of a
production image, rather then something for development.

Here's whats changed:
- Setup Svelte-Kit to build a static page.
- Added code to the server to be able to serve the static files from the
  frontend.
- Added configuration to the package.json to bundle the server code.
- Added multi-stage Dockerfile plus .dockerignore to put it all
  together.