Added more client events for proper disconnecting
Before, when clients went away, all other clients weren't notified and
so broken streams and obsolete watchers were displayed.

Here, I fixed this by giving back more authority to the server. I wanted
to ahve a stateless server. But I already compromised this by having to
implement a client map to avoid duplicate connections when developing
with Vite. The HMR causes a new connection to the signalign server to be
established. Also, clients might not reliably be able to communicate
that they are going away. For example, I cannot detect the device going
to sleep in the browser.

Anyway, the server now has the authority to communicate clients
connecting and disconnecting. The latter event I am now using to clean
up dead connections in the ConnectionManager and in the index page.