Gracefully close all connection when page unmounts
In this commit I added a bit of code that allows the index page to close
down all peer connections as well as the websocket connection to the
signaling server when unmounting.

The app technically does not require this at this point, as there are no
other routes or pages. I added it mostly becayse it feels right. But I
also noticed my browser getting slower and slower the more HMR remounts
the index page.