Implemented possible mitigation for broken streams after device sleep
When the sending device is send to sleep, after a while the webcam feed
ends. When waking the device up, now the webcam feed is just black.
Yes, a simple reload would fix that, but I want dead simple convenience
and reliability.

And because I cannot reliably detect sleep and wakeup events in the
browser without delays of spamming intervals, I found that the
MediaStreamTrack, just has an event for when it ends. My hope is, that
this event is fired reliably, after the device wakes up and the
webcamfeed has ended. This saves me the hassle of implementing some
sleep detection and should work much more direct. Plus, if the feed ends
for any other reason, it is directly restarted.

I cannot test this on my local machine, so I need to push this and see
how it behaves in "production".