Added connection indicator
In order to be transparent about the state of the connection of peers, I
added a colorful indicator. Because my wife kept reloading the page to
make sure the feed really was live.

I decided to differentiate three states:
Connected, Error, and Connecting. All disconnected states (failed,
disconnected, closed) are regarded as an error, as the connection needs
to be permanent. The state 'new' is regarded as connecting as well.

Here's whats changed:
- Added a ConnectionIndicator component, that just renders colorful
  dots. Animated when connecting.
- I implemented the EventSystem into the PeerConnectionHandler and added
  lister for the connectionstatechange event that expose this event to
  outside observers via the EventSystem.
- I added a connection state to the Transmitter and Receiver component,
  making use of the new connectionstatechange event on the
  PeerConnectionHandler. The connection state is then used to render the
  ConnectionIndicator.
- I decided to merge the Receiver and RemoteScreen component. I didn't
  feel like passing all the props to the RemoteScreen component, they
  would be so closely tied together then.