New theme color and font
Pink was nice and baby like, but I want to keep with the yuzu theme, for
upcoming error messages that will include a baby yuzu character. And
pink does not match well with yellow.

I also added a cute font - for japanese characters as well, although
i18n is somewhat far away.