Loading animation for streams
In this commit I added a loading animation for the local camera feed as
well as incoming remote streams.

This way the user sees that there is something coming soon, with the
name of the client already visible.

To me, this makes me more patient with the app I am using. Just a blank
screen gives me the impression something is broken.