Added empty screen for when no cameras are available
When there are no cameras to receive feeds from, the screen used to be
just blank - which I don't consider good ux. I want my software to
communicate with the user.

So in this commit I added some cute image to show up as well as a
message telling you that there are no cameras available and what to do
about that.