Automatically reload app, when server updates
In this commit I implemented an automatic relaod, whenever a version
misatch between the server and the client is detected. This makes
updating the app even more handsoff - ease of use, just what I want!

I separated the SignalingSocket creation from the peer management logic.
The socket is now created in the layout - does not feel super right, but
also not completely wrong. Making this into a component also didn't feel
right.

So, in the layout I am now creating the SignalingSocket, add listeners
for to know when to reload and when the connection is established.

When creating the SignalingSocket, a versio must now be provided. This
is used by the server to automatically close the connection. I am doing
this in the WebSocket part, as this way I can set a code, to give a
reason why the connection was closed that I then can access on the
client. Not upgrading the connection and returning a https status code
does not work as there is no way to get the http status code from the
WebSocket object or event.

I added an enum for keeping all the reasons and codes for why a
connection got closed together and also give them a human readable name.