Fixed and improved responsiveness
Here I fixed and improved the mobile layout of the app. Video feeds
are no longer overflowing the viewport, I am using 100svh instead of
100vh to accomodate any urls bars on mobile, I also separated the
message component and the video feeds so that they don't share any
parent components and are free to fill the screen as they please, this
just makes styling them easier without less breaking. Lastly I decided
to reduce the padding around the video feeds to use more of the screen.