Licensing and Readme
Added licensing in reuse compatible format. App it MPL-2.0 while the
assets are CC-BY-NC-4.0. This decision was made to protect this app from
being resold without any significant or value-adding changes. And also
just as an excercise for me learning how licensing and reuse works.

I also updated the readme to actually contain information about the app,
guidance on how to set it up and licensing information.