Mockup for visual explanation
Here, I created a mockup of a couple devices describing the basics of
how to use the app.

I decided to create the mockup in HTML as this is more reuseable that
creating it in GIMP. That is also the reason why all those new assets
are checked in.

The route for the mockup, I unfortunately could not exclude from the
build with svelte. While it would be nice to not have this route in, it
also does no harm. So I'll leave it in for the moment.

I updated the readme, to include the mockup.