Migrated BatteryServie to shared rune
Just as with the camera permissions, I also migrated the battery state
to a shared rune to reduce complexity in components - did not reduce the
complexity meaninngfully in this case, but consistency is important.

Also, in doing so, the battery state reading should be more robust and
not break in browsers that do not support it.